fix(checkbox): call onChange only once and dont on "set checked"
The keypoint here is, that this PR basically fixes a very old behavior in SUI:
According to the docs the behaviors `set checked|unchecked|enabled|disabled|determinate|indeterminate` are not supposed to run callbacks. But they actually always did (!) They ever since triggered the `change event`, but it never had any impact, just because the module itself did not have any `onChange` event until #295
That said, in conclusion the onchange event was now called twice: In the `set` methods aswell as in the `module` methods. :disappointed:  The onChange event has to be kept, otherwise the keynav will not notify any change to the event system, so i removed the (wrong) event triggering in the above mentioned behavior functions and put them outside instead.

In addition i also hotfixed #295, because the beforeChecked/Unchecked methods were not called correctly 


Closes #399, #400