feat(dropdown, search): support to ignore diacritics on search
This PR adds the option `ignoreDiacritics` (default `false` to stay backward compatible) to the dropdown and search module.
When activated, searches will also match results for base diacritic letters. For example when searching for 'a', it will also match 'á' or 'â' or 'å' and so on...
It will also ignore diacritics for the searchterm, so if searching for 'ó', it will match 'ó', but also 'o', 'ô' or 'õ' and so on...

For the dropdown module i also optimized the sorting method to sort `<option>` values to make sure Diacritics Values are ignored in sorted order. See screenshots below

I implemented the easiest, shortest and most effective solution by using the `String().normalize()` method.
As unfortunately IE does not support this, the modules will check support at initialization time and reset the `ignoreDiacritics` setting accordingly, so it behaves as before in such case.
There exists a polyfill to also support IE, so this will be mentioned in the console if needed.