feat(modal): add overlay fullscreen modal
The current `fullscreen` modal was (according to the docs) supposed to "use the entire size of the screen". This was not really true, because there still was lots of margin left, so you could see the dimmer, and also modals with little content were not even close to fullscreen.

This PR makes it possible now and adds a new variant to the modal module:

`overlay fullscreen modal`

When used, it is now **really** using and overlaying the entire viewport, regardless of the content. 🙂 

Closes #522