fix(dropdown): click-event was locked when single selection had allowAdditions enabled
When a single selection dropdown allowed additions which are not available in the dropdown itself by configuring the dropdown using `allowAdditions:true` and such non-existing value was entered, it was not possible to click anywhere else in the document anymore, because the click-event was swallowed.

Reason for this was that the dropdownmenu was not active but still visible. The empty dropdown was not animated to be hidden anymore and the unbinding of the click event  (which prevented any other click event)  did not take place, keeping all following clicks for the whole document in a trap. 

Closes #591