fix(form): Remove unnecessary property '-webkit-appearance: none' with :focus from error field
The property '-webkit-appearance: none' causes the select element unstable on focus and unfocus
events. Removing native appearance from select on focus reduces the space between the border
and it's content which causes unexpected bouncing.

This property doesn't give any extra effect to all fields for focus event, but making select
element inconsistency alone. So, better to remove this property to make the select element stable
along with the other fields when it's focused or unfocused.