[Card] Horizontal cards meta, extra content and actions are aligned properly
Mata data, actions or buttons and extra content are forced to display in one
row because of 'flex-direction: row' setting 'flex-wrap: wrap', and meta data
flex item has no initial size. So, it forces the extra content and buttons
to take available space next to it which leads them displaying in the same row.

Since, the extra content and buttons are always supposed to appear on the next
line and should take 100% width of the card. The meta data should have initial
size with minimum value which will make sure the meta data to take the available
space on the current row and the extra content and buttons to appear in the next
row respectively.