feat(popup): add optional setting 'forcePosition' to disable calculation
This PR adds a new optional setting forcePosition (default false to stay backward compatible) to the popup module.
Before this PR the popup module always tries to recalculate a given position setting, in case the popup wouldn't be completely visible. if now forcePosition: true is used, the position setting will definately be used, regardless if it's completely shown or cropped. This could help in situations for the calendar module because the (non-inline) calendar is a popup itself and disturbs wanted UI behavior.

Closes #684