fix(accordion): remove transition class after animation so loaders work
A loading transition moves into invisible area, which is correct behavior.
However when some content, like a form or segment, got the transition class after getting animated and you try to add the loading class onto it (for example to show a loading form or segment), the loader and whole content vanishes, because the transition class is still set and prevents the loading display.
Exactly that happens in the accordion module when child elements are getting animated. They keep the transition class.
This PR now removes the transition class after show-animation finishes (only there! hidden transition must be kept!) to prepare the element for usage of a loading class again. This is a safe action, because any additional animation (opening/closing accordion) will always re-add the transition class again because of the usage of the transition module and on initialization the opened accordion does not have any transition class at all (thats why it always works the first time)

Closes #865