fix(modal): vertical alignment on detachable modal was incorrect
When modal is not detachable, it use the legacy display system, not flex based. So it calculates the position of the modal based on his window position, then apply a margin to center it. It works, when you don't have scrolled...

For me, the margin trick is not the best solution, so i tried to center the modal by calculating the top property instead (hope what I write is understandable), but with including the scroll position.

This is not a breaking change, but regarding the nature of the bug I think that we need to be careful with this PR.

Et voilà !

Closes #1009 #717