fix(form): trigger correct dirty handlers to fix function of checkboxes
This PR fixes a behavior when non JS initialized components were used in forms. Especially checkboxes were affected.
The dirty-events were triggered on appropriate change click keyup keydown blur events.
By using jQuery's .trigger() method it was not only triggering the wanted .dirty events, but also any existing behavior having the same name. Unfortunately it exists jquery's .click method, which was now also called whenever a raw (non JS initialized) component like checkbox was clicked.
This resulted in un-clicking the checkbox immediatly when it was clicked. Thus clicking on it had no effect.
Whenever a form component already was initialized by JS before, this was not happening, because there the click-event was explicitely handled in the appropriate module...

The final fix was to use jquery's .triggerHandler() instead, which makes sure only the events are called and same-named bahaviors are ignored.

Closes #1053