fix(dropdown): parameter count resulted in wrong data for onchange
BREAKING CHANGE: When in multiple selection dropdowns a value is selected, the amount of parameters provided to the set value method is not matching. It provides 4 values instead of 3 expected, therefore inconsistent, and because of this also in wrong order. Four parameters are never used anywhere else and not even documented

Providing the additional single value is not necessary because there also exists the onAdd event fetching that value and in multiple dropdowns its only the value itself that counts.