feat(calendar): add onBeforeChange callback and fix a date value issue
BREAKING CHANGE: This PR actually provide two things:

a new callback, called onBeforeChange, which allow to cancel a change before it's applied to calendar logic. The old onChange callback is now moved further.

a fix to the onChange callback. His position was bad, since the date wasn't really changed internally when it was called, so any call the one of the behaviors or using the input resulted in a bad value.

⚠️ This is a breaking change ! The new onChange callback, doesn't allow to cancel change anymore, so we might warn users when we'll publish this update.

Closes #417