fix(popup): avoid detection of wrong existing popup
When a popup was defined to get its content out of an existing DOM-Node it was in some situations wrongly reused on another, totally different popup because the popup-creation code was searching for existing nodes _before__checking for a specific domnode. That way a previously detached domnode from another popup was used by accident.
This PR switched priorities to always check for a specific defined DOM node in settings.popup before looking for an existing dom node next to the popup.