fix(dropdown): remove required attribute from select
When a dropdown was initialized out of a select tag and that tag had the 'required' attribute set, it is not possible to submit the form with an empty value, because the select-tag stays hidden. Instead a console warning is shown "An invalid form control with name='' is not focusable"

Basically we could argue "Just don't add 'required' to the select tag and use FUI form validation instead". But there might be situations where the form is generated by some backend framework where the "required" attribute is set automatically, thus i just removed the required attribute in any case as a proposal to solve the situation.

This PR possibly breaks existing code, in case some developer have used the "required" attribute explicitly to gain the current behavior (and don't care about the console warning)

Closes
#917