fix(modal): basic modal header should not have a border
A Modal header has a bottom border. A basic modal instead should not ( the border for actions in basic modals is already removed, and it just looks misplaced, especially when there is noticable background)

Using the default dimmer settings the dimmer color is identical to the bottom border, thus invisible.
However, when the dimmer has a different shade or is inverted, the border gets visible.