feat(form): auto add empty rule validation for required fields
This PR adds the ability to add automatically an 'empty' rule for fields that got the required class (or checked for checkboxes). It's disabled by default since I don't want it to be a breaking change, but can be enabled at initialization with autoCheckRequired: true or any time with the .form('set auto check') behavior.

It can handle all fields types, grouped fields as well, and is able to detect disabled fields or groups. Rule is added to others if there's already some defined, and isn't duplicated if already set.