fix(table): invented a helper class to solve the styling logic when rowspan is used in tables
The styling is sometimes broken in the table element, when rowspan used. This confused later table rows which then usually have less columns.

As there is no way to look backward via CSS to find out if a previous column in a row was rowspanned and the table component is not a JS module, i invented a little helper class to solve this issue via pure html/css.

One just has to create a td node , where it is usually not necessary, and give it the new rowspanned class. This will just hide it from the screen but all existing table classes are working again, because the amount of td remains the same just like if there was no rowspan used at all.