fix(build): prevent gulp watch task from refreshing all components
When defining individuals inside semantic.json, these components are added to the list of regular components inside the build function.

The problem is that this build function is also used by the 'watch' task to refresh individual files. So when running gulp watch with individuals defined, the src variable containing the path to the watched file is overwritten by the full list of components. The result is that each change triggers a full rebuild.

Not what we want, so I added an extra condition to the if statement that's changing the src variable. It seems the src variable is always passed as object when using watch.. So by checking if src is a string, it is only altered during regular build tasks.