fix(calendar): inputClick adds on existing popup events
This commit should fix an issue where a popup would be called to be shown twice if the calendar was bound to an input field. The popup module is already adding these events to inputs.

This can be a big performance issue on bigger DOM trees as the popup position calculation is very heavy.