Split tertiaryLinePadding variable
I wanted to set a horizontal padding of 0 for tertiary buttons, so I changed tertiaryLinePadding to '0.5em 0'. But it resulted in a build error, because this variable is also used for calculating tertiary button margin.

I think it's better so split this variable into vertical and horizontal values and then use the vertical value for calculating the margins.

This also aligns better with other button variables (verticalPadding and compactVerticalPadding).