fix(calendar): fix the popup show function getting called twice
This PR fixes an issue where the popup would be called to be shown twice because the two modules are listening to the events to show the popup when only one should be. I've thus removed the calendar's listeners since the popup ones that should be expected.

This PR also fixes the end calendars' popup which were also getting called to be shown twice when using the default settings for input calendars. The calendar module was focusing the input which is triggering the focus event that is also listened by the popup module. With this PR, the popup's show method is called only if the listener is not on focus, so that the input can still be focused.

This can be a big performance issue on heavy DOM trees as the popup position calculation is very heavy.