fix(dropdown): nested dropdown positioning
When the dropdown is used as simple dropdown without JavaScript,
it doesn't support more than two nested level of sub menus. It
always exposes only two levels of sub menus.

And despite the sub menu declares the value 'left' for the class
name, it always apear rightward. It should respect the value
'left' from the class name and should be appeared leftward.

When the dropdown menu is invoked with JavaScript, the last dropdown
in the right menu doesn't behave properly and the sub menu inside
the dropdown are overlapping each others. They should appear in the
correct location like other dropdown menus.