fix(tab): make auto activation of tabs optional
The implemented auto activation of tabs when no active tab is found lead into some situations where a manual set of the active tab does not work anymore and always the last tab was selected.
This especially happens when tabs are initialized separately. Such situations need the old until FUI 2.7.7 which did not check for existing active tabs at all

This PR adds a new option autoTabActivation (default true to avoid being a breaking change within the 2.8.x branch)
When set to false the behavior is the same as until 2.7.7
When set to a string and no active tab exists, it will use that string as the path name for the tab to be activated.