fix(dimmer): page dimmer in firefox results in misaligned modals
Firefox messes around with (large) modals in certain situations. For example when trying to drag inside an input inside a modal, it scrolls to the top of the modal. Even worse on Mac where the whole modals overlays other div containers...

Reason for this is the perspective setting for the page dimmer. It seems this was introduced some years back to also fix some firefox issues, so i don't dare to remove this setting completely to avoid possible other, currently unknown, issues in probably other browsers as well.