fix(dropdown): fix some bug introduces by the previous commit (#1473)
This PR fixes the first level of simple dropdown inside the right menu
was cropped and out of the screen when the parent menu item has short
text, and the dropdown doesn't specify the class name `left`.

To avoid this, all dropdown inside the right menu will be forced to
appear leftward by default. The dropdown can also appear leftward or
rightward by having `left` or `right` class name.

This PR also fix the dropdown inside the buttons which are positioned
incorrectly.

Closes: #1502