fix(multiple): make sure real icon selectors are adressed
In certain situations only .icon selectors were used to address icons incomponents. However this interferes with a possible icon button resulting in wrong sizes or wrong aligned margins and colors for those buttons.
This PR adds the i tag to those selectors to make sure only a "real" icon is selected as it was supposed to be.