fix(calendar): adjust tooltip positions according to popup  position
The default tooltip position of cell tooltips is left center. This leads into situations where the tooltip is cut to the left.
On first instantiation of a calendar popup the position of tooltips is not correctly fetched because the popup module hasn't adjusted its pointer position classes yet. Because of this ,the direction of possible cell tooltips is wrongly calculated.
If the same calendar is opened a second time, this does not occur anymore, because the popup module has now adjusted its position.

This PR now adjusts any existing cell tooltip positions inside the calendar whenever the popup get's shown (the popup module always (re-)calculates its position then). So this not only fixes it on first instantiation but also in case the calendar position changes while scrolling , so the popup will possibly to displayed in another direction.

As this is not a module setting but a decision of the module itself, this is a backward compatible change which just fixes the behavior.
Btw. it wasn't happening before, because the calendar popup was called twice (thus correcting it's position), but that was now fixed by #1439