fix(image): support centered div and adjusted aligned groups
The current image component can only be centered with img element. It is not able
to center when using a different HTML element, for example, an image link with
anchor tag or a DIV container.

This PR would support to center the image not only for img tag, but also for other
HTML element by using `centered` flag.

This PR would also support to display the group images horizontally center. Additionally
the group images can now share the same vertical alignment as `top`, `middle` or `bottom`
together or can have it's own vertical alignment.

Closes: #1608