fix(sidebar): avoid scrollbar on non body pushable
When a bottom sidebar was used inside non body pushable context, it was always added to the context divs height, making the vertical scrollbar appear. As the body will still provide a possible scrollbar it's save to remove the inner pushable vertical scrollbar.

Reason for the behavior is the 100% translate3d() setting to hide the sidebar which adds its content to the parent div. For left/right the overflow-x:hidden is already set and for top it's a negative position of -100%, so the change of overflow-y is not affected in the other cases.