fix(form): isvalid should not delete error state and made optional now
the is valid behavior deletes possible existing error states on a field, although it shouldn't.
That's because the field validator always removes the error state by default (before probably adding it again) and did not respect a showErrors setting. showErrors however is only meant to show new errors. If it is false it should not touch any existing error state class.

I now also added the showErrors parameter as an option to the is valid behavior in case someone definately wants to show errors on not already validated fields.

As the parameter is optional the change is fully backwards compatible.