feat(progress): total value of 0 was unsupported
A total value of 0 was unsupported and wrongly fetched (0 !== undefined !)
Although this might be a rarely used situation, it should behave correctly. Means an explicitely given total value of 0 is always fullfilled and should behave as 100%
This PR also fixes wrong handling of the limitValues setting and a wrong check in precentage calculation for the total value which might be undefined and resulted in NaN numbers otherwise.