feat(dropdown): support separate transition displaytype
In situations where a dropdown was inside a right menu (which got a flex setting) the transition module was taking that flex displaytype setting for the dropdown menu. This unfortunately results into a horizontal menu list.
To be able to adjust this, i added the displayType, used for the transition, setting to the dropdown module as well. If set, this setting could be used just as in transition to correct the display if needed.

Additionally this setting also had to be provided to the hiding transition, otherwise the wrong flex setting would be reapplied before hiding the dropdown again.