feat(dropdown): support preventChangeTrigger parameter for set.selected
This PR simply add the ability to prevent the onChange event trigerring when the set selected behavior is used. Until now, the only way to not trigger the onChange was to use the set value behavior, which is not IMHO the right stuff to call when you want to programmatically select a value.