fix(dropdown): fix menu show on mobile and api call
When a search dropdown was used, loading remote content via apiSettings, the menu did not show up on mobile devices.
This was due to the lost search focus which happened when the menu nodes were recreated after the api returned the dropdown items. Mobile devices triggered a blur of the search field, while desktop browsers did not. Blurring the search field however led to not showing the dropdown menu anymore.

This PR also fixes that the API was called twice the first time one clicked on the search field (while it was only triggered once when clicking on the dropdown icon)