[Table] Allow pointer events for the first header cell in definition table
The first header cell of definition table disables the pointer event which
causes the problem when the cell contains the content that are needed the
pointer event such as dropdown, input field or the content from title
attribute is unable to be seen.

This PR changes the behavior of pointer event to become available if there
is content inside the cell.

Closed: #1756