fix(calendar): return null instead of undefined when invalid date is passed to date sanitizer
The date sanitizer returns undefined when the date argument is not a valid date object or
date string which doesn't allow to reset the date value especially for min and max date by
passing null value via `set minDate` and `set MaxDate`.

This PR makes the `sanitiseDate` method to return null if the value of date argument is not
valid date object or date string.

Close: #1776