fix(dropdown): calculate correct height of current menu if it contains child menu
When the menu contain the child menu, the calculation of menu height is incorrect
while evaluating to open upward or downard.

It's because the child menu is hidden when the current menu is in the transition of
being open and it doesn't occupy any space in the current menu and the height of
the current menu is incorrectly calculated without the height of child menu which
let the menu opens upward or downward incorrectly.

This PR checks if the current menu has the child menu and adds the child menu height
to the current menu height to give correct assumption whether the menu fits in the
screen upward or downward.

Close: #1711