fix(dropdown): Clearing multiple selection dropdown supports to prevent onChange callback
When clearing the selected items from multiple dropdown via `clear` method, the
`onChange` callback triggers although the `preventChange` is set as true.

It's because, when the dropdown is multiple selection, the `clear` calls the
specific internal method to remove multiple selected items without fowarding the
`preventChange` setting which results triggering the `onChange` callback.

This PR makes the `clear` method to call the internal method and pass the `preventChange`
setting which supports to prevent triggering the `onChange` callback.

Close: #1781