fix(build): use current year and implement version in header
This PR adjust some build settings. It...

prepares for the next 2.8.8 release by updating the version number in package.json. For 2.8.7 this was done after the NPM upload, so the 2.8.7 release had the 2.8.6 header.
fixes inclusion of the version number in the comment headers of each file (was omitted since a few versions)
fixes to adjust the year in the comment header to the year of release (was a hardcoded 2014 before)