fix(modal): bodyfixed selector did not work with existing padding
When a modal opens and the screen ist large, so there exists a scrollbar, selected fixed elements on the screen are adjusted so they dont move while the screen is shown/closed.

The fixed nag element was missing to be supported
existing values for padding were not taken into account. For all previous selectors (toast-container, menu, sidebar) this was not a problem, because they haven't had a padding by default. But the new nag selector actually has a padding of 1em.
In case a non fixed selector was used, the padding was the wrong approach (for example the close icon inside the nag)
When the modal hides, the previously set values for padding were overridden by 0. Whenever there was a default padding before, this would not work anymore. I now completely removed the inline style instead.