fix(dropdown): fix menu item cannot break the text inside
When the menu item contains too long text, the text doesn't break
into the new line when there's no enought space to display in
single line and the text overflows the container.

This PR fixes the menu item text breaking problem and also changes
the container to flex box to render the image, text and the dropdown
icon as flex box items to align better.