feat(icon): colored variant for bordered or circular
This PR adds a colored variant to circular or bordered icon(s).

Currently the border of a circular or bordered icon is always grey. Even if the icon itself has got a color.
By adding a colored class to such icons the border will inherit its color from the icon itself.