fix(calendar): blur input on hidden
This PR just blur the calendar when a date is selected and the popup has closed. Before, selecting a date just hide the popup but it still was focused. Now if you select a date the popup will hide and the input will be blurred. So you can click again to the input to reshow the popup.