fix(toast): dedicated horizontal container
Whenever a horizontal toast was created, all further toast, in the same container position, were shown horizontal as well, even if they were supposed to be displayed normal/vertically.

Reason for this was the container selector, which was only defined for the positioning, but ignored a possible horizontal setting. This PR now creates a dedicated horizontal container, which fixes such situations.