fix(form): avoid double submit on enterkey
Whenever the enter key is pressed inside a form field, it triggers a form submit. However, this is also a browser default behavior, thus the event was called twice. Especially for IE11 this was causing the normal submit to not function anymore.
I also made sure that native input fields of type "reset", "button" or "submit" are not fetched for input validation anymore (just in case someone would try). A native "reset" button would otherwise also submit the form (when enter is pressed on focussed button)