feat(modal): a11y support, tabbing in modal, close icon focus/keypress
This PR adds accessibility (a11y) support to the modal module

add proper aria attributes
make sure tabbing only takes place inside the modal and does not tab to fields hidden by the dimmer
make a close button a button and tabbable as well
I also separated the closeIcon from the previous close selector which was supposed to be able to also close the modal on each action button when the selector is enhanced.
Reason for separation of closeIcon is the new ability to trigger the closeicon by space or enter which is needed as accessibility requirement and was not possible before.

Attention
The aria attributes are only applied when modal is called dynamically via JS attributes (means without an already prepared modal DOM structure). When using existing Modal DOM nodes, adding proper aria attributes/labels is still the job of the developer as Modal is not adjusting existing DOM node attributes (for example an id has to be created for aria-labelledby).

However, the tabbing feature works independently of existing aria labels and was a general bug anyway.