feat(checkbox): detect on as checked and add checkbox string values
Using set values for a checkbox field inside a form will only accept "true" or "1" to recognize the field as checked.
However, the standard value for a checked checkbox (when retrieved by "get values") is "on".

This PR now accepts a possible "on" value for given checkbox fields to also check the checkbox to have the same bahavior.

In addition, any string value will now be set as value for the checkbox as well. (This is valid specification as of developer.mozilla.org/en-US/docs/Web/HTML/Element/Input/checkbox#value), but any other string than "on" will not check the checkbox.