fix(grid): stretched rows were broken on tablet size
A stretched row inside doubling grid was not stretched on tablet size.

Yes, i left the !important in because there are many grid selector combinations where it probably breaks other situations which would need much more investigation. As the !important was there before, i decided to leave it in for now.