fix(icon): adjust icons width to match single icon
I have a dropdown with two items

The first item has a simple icon

<i class="blue file excel outline icon"></i>

The second one has two grouped icons

<i class="icons">
    <i class="red file excel outline icon"></i>
    <i class="red bottom right corner key icon"></i>
</i>
Because the .icon has a width and the .icons does not visually they look off. The grouped icons has a smaller width then the simple icon

This PR adds a min-width to the .icons to make sure that the .icons group will be at least the same size of the .icon
I didn't want to use width because in some cases maybe the grouped icons need more space.

For the same reason i added a min-width i added also a min-height.
the height is not the cause of my problem, but i think it's best to add it as well.