Fix exclusion of table component when tab is active
If there is still a table.css present in dist/components, it will be included in semantic.css.

This is caused by a collision in the globbing mechanism: 'tab' is included in the word 'table', so table.css is included in the package task.

Fixes https://github.com/fomantic/Fomantic-UI/issues/2118