feat(chore): color list for each component separately
This PR allows to define a color list for each component individually out of the central colors.less file.

For example, now it is possible to have only red, blue and orange buttons, but only yellow labels and no colors for message at all. See example below.

All of this can be defined in the central variation.variables file as before (or in custom theme file)

Before this change you were only able to reduce the whole color list by modifying the colors.less file and change the available colors for all components at once.