fix(modal): outside close button handling
when you have two modals on top of each other, if the height of both of them are the same and close buttons are out of the modal, the close buttons are on each other which has a bad user experience.

In addition the close icon from the first modal was still clickable and would close the first modal.

In some cases the close icon could be shown over a different inner Dimmer making the close icon invisible, so i added a slight text shadow to keep it visible. This text shadow is not recognizable on usual modal usage .

At last i made the dimmer class of the inner dimmer optional, so one could customize that as well.